refactor: extract Report service into Action and Query classes
Decompose ReportBuilder into GenerateReport action and GetReportPreview
query under Domain/Report. GenerateReport uses aggregate root methods
(addDay, addTask). Update ReportController to method injection. Delete
ReportBuilder, its interface, and update ReportServiceProvider.
