+++ backend/app/Providers/ReportServiceProvider.php
@@ -6,8 +6,6 @@ namespace App\Providers;
 
 use App\Services\Report\PromptExportService;
 use App\Services\Report\PromptExportServiceInterface;
-use App\Services\Report\ReportBuilder;
-use App\Services\Report\ReportBuilderInterface;
 use App\Services\Report\ReportExporterInterface;
 use App\Services\Report\WordExporter;
 use Illuminate\Support\ServiceProvider;
@@ -16,7 +14,6 @@ class ReportServiceProvider extends ServiceProvider
 {
     public function register(): void
     {
-        $this->app->bind(ReportBuilderInterface::class, ReportBuilder::class);
         $this->app->bind(ReportExporterInterface::class, WordExporter::class);
         $this->app->bind(PromptExportServiceInterface::class, PromptExportService::class);
     }
